requirements.txt
requests
python-telegram-bot
telethon
ntscraper
python-dotenv